make the singularity check in 8-point method less strict.
git-svn-id: file://localhost/home/tolik/svnrepo/trunk@1302 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/src/cv/cvfundam.cpp
@@ -860,7 +860,7 @@ int CvFMEstimator::run8Point( const CvMat* _m1, const CvMat* _m2, CvMat* _fmatri
 
     for( i = 0; i < 8; i++ )
     {
-        if( fabs(w[i]) < FLT_EPSILON )
+        if( fabs(w[i]) < DBL_EPSILON )
             break;
     }
